Do not deploy web site as part of firmware.
The site is getting too big to server effectively, so it will now be
loaded from another dedicated webserver which will allow the Hexapod API
host address to be supplied as a site input.
The site will allow CORS for any external host, so should be able call
the Hexapod API even if not served from the same host.